Set type, Key paths, Dictionary optimisation

+++ Source/KVS/kvHash.pas
@@ -86,7 +86,7 @@ begin
 
       F := D xor (D shl 16) xor (D shl 32) xor (D shl 48);
       H := H xor F;
-      F := L xor (Int64(I) shl 18) xor (Int64(L) shl 28) xor (Int64(I) shl 31);
+      F := L xor (UInt64(I) shl 18) xor (UInt64(L) shl 28) xor (UInt64(I) shl 31);
       H := H xor F;
 
       H := TransformHash(H);
@@ -95,11 +95,11 @@ begin
       H2 := Word32(H and $FFFFFFFF);
       H1 := Word32(H1 + Word32(I) + Word32(D));
       H2 := Word32(H2 + Word32(I) + Word32(L));
-      H1 := Word32(Int64(H1) * 73 + 1);
-      H2 := Word32(Int64(H2) * 5 + 79);
+      H1 := Word32(UInt64(H1) * 73 + 1);
+      H2 := Word32(UInt64(H2) * 5 + 79);
 
-      T1 := Word32(H1 shl 11) xor (H1 shr 5) xor Word32(H2 shl 17) xor (H2 shr 19);
-      T2 := Word32(H2 shl 7)  xor (H2 shr 3) xor Word32(H1 shl 15) xor (H1 shr 17);
+      T1 := Word32(UInt64(H1) shl 11) xor (H1 shr 5) xor Word32(UInt64(H2) shl 17) xor (H2 shr 19);
+      T2 := Word32(UInt64(H2) shl 7)  xor (H2 shr 3) xor Word32(UInt64(H1) shl 15) xor (H1 shr 17);
       H := (UInt64(T1) shl 32) or T2;
     end;
